Pare down moved over code
