--- src/applications/personalization/dashboard/actions/messaging/index.js
@@ -1,13 +1,0 @@
-import { UPDATE_ROUTE } from '../../utils/constants';
-
-export * from './alert';
-export * from './compose';
-export * from './folders';
-export * from './messages';
-export * from './modals';
-export * from './preferences';
-export * from './search';
-
-export function updateRoute(location) {
-  return { type: UPDATE_ROUTE, location };
-}
