+++ src/applications/personalization/dashboard/actions/prescriptions.js
@@ -1,32 +1,4 @@
-import recordEvent from '../../../../platform/monitoring/record-event';
 import { apiRequest } from '../utils/helpers';
-
-export function loadPrescription(id) {
-  return dispatch => {
-    const urls = [`/${id}`, `/${id}/trackings`];
-    const errorHandler = () => dispatch({ type: 'LOAD_PRESCRIPTION_FAILURE' });
-
-    dispatch({ type: 'LOADING_DETAIL' });
-
-    // Fetch both the prescription and its tracking history and
-    // wait for retrieval and read of both resources to resolve.
-    Promise.all(
-      urls.map(url =>
-        apiRequest(url, null, response => response, errorHandler),
-      ),
-    )
-      .then(data =>
-        dispatch({
-          type: 'LOAD_PRESCRIPTION_SUCCESS',
-          data: {
-            rx: data[0].data,
-            trackings: data[1].data,
-          },
-        }),
-      )
-      .catch(errorHandler);
-  };
-}
 
 export function loadPrescriptions(options) {
   let url = '/';
@@ -75,40 +47,4 @@ export function loadPrescriptions(options) {
         }),
     );
   };
-}
-
-export function refillPrescription(prescription) {
-  if (prescription.prescriptionId) {
-    const url = `/${prescription.prescriptionId}/refill`;
-
-    recordEvent({
-      event: 'rx-confirm-refill',
-    });
-
-    return dispatch => {
-      dispatch({ type: 'REFILL_SUBMITTED' });
-
-      apiRequest(
-        url,
-        { method: 'PATCH' },
-        () =>
-          dispatch({
-            type: 'REFILL_SUCCESS',
-            prescription,
-          }),
-        response =>
-          dispatch({
-            type: 'REFILL_FAILURE',
-            errors: response.errors,
-            prescription,
-          }),
-      );
-    };
-  }
-
-  return { type: 'REFILL_FAILURE' };
-}
-
-export function sortPrescriptions(sort, order = 'ASC') {
-  return { type: 'SORT_PRESCRIPTIONS', sort, order };
 }
