+++ src/applications/personalization/dashboard/tests/reducers/prescriptions.unit.spec.js
@@ -19,23 +19,6 @@ describe('prescriptions reducer', () => {
     expect(state.history.loading).to.be.true;
   });
 
-  it('should show a loading screen for detail', () => {
-    const state = prescriptionsReducer(
-      { detail: { loading: false } },
-      { type: 'LOADING_DETAIL' },
-    );
-    expect(state.detail.loading).to.be.true;
-  });
-
-  it('should handle failure to fetch a prescription', () => {
-    const state = prescriptionsReducer(
-      { currentItem: { id: 123 } },
-      { type: 'LOAD_PRESCRIPTION_FAILURE' },
-    );
-    expect(state.currentItem).to.be.null;
-    expect(state.detail.loading).to.be.false;
-  });
-
   it('should handle failure to fetch active prescriptions', () => {
     const state = prescriptionsReducer(
       {
@@ -113,146 +96,5 @@ describe('prescriptions reducer', () => {
     expect(state.history.page).to.equal(1);
     expect(state.history.pages).to.equal(1);
     expect(state.history.loading).to.be.false;
-  });
-
-  it('should handle a successful refill', () => {
-    const state = {
-      items: [
-        {
-          id: 1,
-          attributes: {
-            refillStatus: 'active',
-            isRefillable: true,
-          },
-        },
-        {
-          id: 2,
-          attributes: {
-            refillStatus: 'active',
-            isRefillable: true,
-          },
-        },
-        {
-          id: 3,
-          attributes: {
-            refillStatus: 'active',
-            isRefillable: true,
-          },
-        },
-      ],
-    };
-
-    const newState = prescriptionsReducer(state, {
-      type: 'REFILL_SUCCESS',
-      prescription: { prescriptionId: 2 },
-    });
-
-    expect(newState.items[1].attributes.isRefillable).to.be.false;
-    expect(newState.items[1].attributes.refillStatus).to.eql('submitted');
-  });
-
-  it('should sort active prescriptions by prescription name', () => {
-    const state = {
-      items: [
-        {
-          id: 1,
-          attributes: {
-            prescriptionName: 'c',
-          },
-        },
-        {
-          id: 2,
-          attributes: {
-            prescriptionName: 'a',
-          },
-        },
-        {
-          id: 3,
-          attributes: {
-            prescriptionName: 'b',
-          },
-        },
-      ],
-    };
-
-    const newState = prescriptionsReducer(state, {
-      type: 'SORT_PRESCRIPTIONS',
-      sort: 'prescriptionName',
-      order: 'ASC',
-    });
-
-    expect(newState.items[0].id).to.equal(2);
-    expect(newState.items[1].id).to.equal(3);
-    expect(newState.items[2].id).to.equal(1);
-  });
-
-  it('should sort active prescriptions by facility name', () => {
-    const state = {
-      items: [
-        {
-          id: 1,
-          attributes: {
-            facilityName: 'c',
-          },
-        },
-        {
-          id: 2,
-          attributes: {
-            facilityName: 'b',
-          },
-        },
-        {
-          id: 3,
-          attributes: {
-            facilityName: 'a',
-          },
-        },
-      ],
-    };
-
-    const newState = prescriptionsReducer(state, {
-      type: 'SORT_PRESCRIPTIONS',
-      sort: 'facilityName',
-      order: 'ASC',
-    });
-
-    expect(newState.items[0].id).to.equal(3);
-    expect(newState.items[1].id).to.equal(2);
-    expect(newState.items[2].id).to.equal(1);
-  });
-
-  it('should sort active prescriptions by last submit date', () => {
-    const state = {
-      items: [
-        {
-          id: 1,
-          attributes: {
-            refillSubmitDate: 2,
-          },
-        },
-        {
-          id: 2,
-          attributes: {
-            refillSubmitDate: 3,
-          },
-        },
-        {
-          id: 3,
-          attributes: {
-            refillSubmitDate: 1,
-          },
-        },
-      ],
-    };
-
-    const newState = prescriptionsReducer(state, {
-      type: 'SORT_PRESCRIPTIONS',
-      sort: 'lastSubmitDate',
-      order: 'ASC',
-    });
-
-    expect(newState.items[0].id).to.equal(3);
-    expect(newState.items[1].id).to.equal(1);
-    expect(newState.items[2].id).to.equal(2);
   });
 });
